pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-assistant"
version = "0.1.0"
description = "Local RAG assistant (LangChain + Chroma + Ollama)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain==0.2.16",
  "langchain-community==0.2.16",
  "chromadb==0.5.3",
  "sentence-transformers==3.0.1",
  "typer==0.12.3",
  "python-dotenv==1.0.1",
  "pydantic>=2.8.2",
  "fastapi==0.115.0",
  "uvicorn==0.30.5",
  "httpx==0.27.2",
  "jinja2==3.1.4",
  "pytest==8.3.2",
  "ruff==0.6.3",
  "black==24.8.0",
]

[project.scripts]
rag = "rag.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
